# services/ai/openapi.json
{"openapi":"3.1.0","info":{"title":"Smart Lost&Found API","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/db/tables":{"get":{"summary":"Db Tables","operationId":"db_tables_db_tables_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/db/health":{"get":{"tags":["health"],"summary":"Db Health","operationId":"db_health_db_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/items":{"post":{"tags":["items"],"summary":"Create Item","operationId":"create_item_items_post","security":[{"HTTPBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ItemCreateIn"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["items"],"summary":"List Items","operationId":"list_items_items_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/ItemStatus"},{"type":"null"}],"title":"Status"}},{"name":"mine","in":"query","required":false,"schema":{"type":"boolean","description":"내가 등록한 것만 보기","default":false,"title":"Mine"},"description":"내가 등록한 것만 보기"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemOut"},"title":"Response List Items Items Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/items/mine":{"get":{"tags":["items"],"summary":"My Items","operationId":"my_items_items_mine_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/ItemStatus"},{"type":"null"}],"title":"Status"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ItemOut"},"title":"Response My Items Items Mine Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/items/stats":{"get":{"tags":["items"],"summary":"Stats","operationId":"stats_items_stats_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"mine","in":"query","required":false,"schema":{"type":"boolean","description":"내가 등록한 것만 집계","default":false,"title":"Mine"},"description":"내가 등록한 것만 집계"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/items/{item_id}":{"get":{"tags":["items"],"summary":"Get Item","operationId":"get_item_items__item_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"integer","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["items"],"summary":"Delete Item","description":"분실물 삭제 (사진도 함께 삭제)","operationId":"delete_item_items__item_id__delete","security":[{"HTTPBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"integer","title":"Item Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["items"],"summary":"Patch Item","description":"분실물 상태 변경 (전체 업데이트)","operationId":"patch_item_items__item_id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"integer","title":"Item Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Payload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/items/{item_id}/status":{"patch":{"tags":["items"],"summary":"Update Status","operationId":"update_status_items__item_id__status_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"item_id","in":"path","required":true,"schema":{"type":"integer","title":"Item Id"}},{"name":"new_status","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ItemStatus"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/items/candidates":{"get":{"tags":["items"],"summary":"Get Candidates For Ai","description":"AI 서버 전용 엔드포인트\n전체 분실물 후보를 AI 서버에 제공\n\nHeaders:\n    X-Admin-Token: dev-internal-secret\n\nQuery Params:\n    status: 분실물 상태 (기본값: STORED)\n    limit: 최대 개수 (기본값: 50)\n\nReturns:\n    candidates: 분실물 목록 (item_id, name, brand, color, category, stored_place, features, photos 등)","operationId":"get_candidates_for_ai_items_candidates_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/ItemStatus"},{"type":"null"}],"default":"STORED","title":"Status"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":50,"title":"Limit"}},{"name":"X-Admin-Token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Admin-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/search":{"post":{"tags":["search"],"summary":"Search Items","description":"자연어 검색 엔드포인트\n현재는 규칙 기반 필터링만 제공, 추후 LLM 서비스와 통합 예정","operationId":"search_items_search_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchQuery"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["search"],"summary":"Search Items Get","description":"GET 방식 검색 (results 페이지에서 사용)\nLLM 기반 유사도 점수 계산 및 Top 5 반환","operationId":"search_items_get_search_get","parameters":[{"name":"q","in":"query","required":true,"schema":{"type":"string","description":"검색 쿼리","title":"Q"},"description":"검색 쿼리"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/claims/mine":{"get":{"tags":["claims"],"summary":"My Claims","operationId":"my_claims_claims_mine_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/claims/":{"post":{"tags":["claims"],"summary":"Create Claim","operationId":"create_claim_claims__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClaimCreateIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/claims/{claim_id}":{"patch":{"tags":["claims"],"summary":"Update Claim Status","operationId":"update_claim_status_claims__claim_id__patch","parameters":[{"name":"claim_id","in":"path","required":true,"schema":{"type":"integer","title":"Claim Id"}},{"name":"status","in":"query","required":true,"schema":{"$ref":"#/components/schemas/ClaimStatus"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stats/overview":{"get":{"tags":["stats"],"summary":"Overview","operationId":"overview_stats_overview_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/activities/":{"get":{"tags":["activities"],"summary":"Recent Activities","operationId":"recent_activities_activities__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/me/items":{"get":{"tags":["me"],"summary":"Get My Items","description":"내가 등록한 분실물 목록","operationId":"get_my_items_me_items_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/ItemStatus"},{"type":"null"}],"title":"Status"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Size"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/me/stats":{"get":{"tags":["me"],"summary":"Get My Stats","description":"내 통계","operationId":"get_my_stats_me_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/me/activities":{"get":{"tags":["me"],"summary":"Get My Activities","description":"내 활동 목록","operationId":"get_my_activities_me_activities_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/me/profile":{"get":{"tags":["me"],"summary":"Get My Profile","description":"내 프로필","operationId":"get_my_profile_me_profile_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"ClaimCreateIn":{"properties":{"item_id":{"type":"integer","title":"Item Id"},"memo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Memo"}},"type":"object","required":["item_id"],"title":"ClaimCreateIn"},"ClaimStatus":{"type":"string","enum":["PENDING","APPROVED","REJECTED"],"title":"ClaimStatus"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ItemCreateIn":{"properties":{"name":{"type":"string","title":"Name"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"brand":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Brand"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"},"color":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Color"},"material":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Material"},"size":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Size"},"features":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Features"},"accessories":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Accessories"},"serial_masked":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Serial Masked"},"lat":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lat"},"lng":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Lng"},"stored_place":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stored Place"},"photos":{"items":{"$ref":"#/components/schemas/PhotoIn"},"type":"array","minItems":2,"title":"Photos"}},"type":"object","required":["name","photos"],"title":"ItemCreateIn"},"ItemOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"color":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Color"},"status":{"$ref":"#/components/schemas/ItemStatus"}},"type":"object","required":["id","name","status"],"title":"ItemOut"},"ItemStatus":{"type":"string","enum":["STORED","CLAIMED","HANDED_OVER"],"title":"ItemStatus"},"LoginIn":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"LoginIn"},"PhotoIn":{"properties":{"url":{"type":"string","title":"Url"},"exif_json":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Exif Json"}},"type":"object","required":["url"],"title":"PhotoIn"},"RegisterIn":{"properties":{"email":{"type":"string","format":"email","title":"Email"},"phone":{"type":"string","title":"Phone"},"password":{"type":"string","title":"Password"},"role":{"$ref":"#/components/schemas/Role","default":"SEEKER"}},"type":"object","required":["email","phone","password"],"title":"RegisterIn"},"Role":{"type":"string","enum":["SEEKER","FINDER","ADMIN"],"title":"Role"},"SearchQuery":{"properties":{"query":{"type":"string","title":"Query"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"color":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Color"},"brand":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Brand"}},"type":"object","required":["query"],"title":"SearchQuery"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}